requirement.txt
cycler==0.10.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
pandas==1.0.0
patsy==0.5.1
pycausalimpact==0.0.16
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
statsmodels==0.11.0
appnope==0.1.0
astroid==2.3.3
astropy==4.0
attrs==19.3.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
confuse==1.0.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
htmlmin==0.1.12
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
kiwisolver==1.1.0
kneed==0.5.1
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml==4.5.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
packaging==20.1
pandas==0.25.3
pandas-profiling==2.4.1
pandocfilters==1.4.2
parso==0.6.0
patsy==0.5.1
pexpect==4.8.0
phik==0.9.8
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
py==1.8.1
pyarrow==0.15.1
pycausalimpact==0.0.16
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
pytest-pylint==0.14.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
statsmodels==0.11.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
zipp==2.1.0
